requirements.txt
streamlit>=1.34.0
pandas>=2.2.2
databricks-sql-connector>=3.5.0
pyvis>=0.3.2
graphviz>=0.20.3
python-dotenv>=1.0.1